cards - Refactor
