--- Dutify/src/components/addCard/addCardStyle.css
@@ -1,36 +1,0 @@
-#main_container {
-  cursor: pointer;
-  aspect-ratio: 16/9;
-  width: 300px;
-  min-width: 175px;
-  overflow: hidden;
-  transition: 0.25s ease;
-  border-radius: 1rem;
-  border-color: gray;
-}
-  
-#main_container:hover {
-  box-shadow: 0 0 1rem rgba(0, 0, 0, 0.5);
-  transform: scale(1.005);
-}
-  
-.overlay{
-  background-color: #676565;
-}
-  
-.addCard-title { 
-  user-select: none;
-  font-weight: 500;
-  letter-spacing: 1.05px;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  overflow: hidden;
-  color: rgb(223, 223, 223);
-  max-width: 90%;
-}
-
-.addButton{
-  width: 45%;
-  height: 45%;
-  color: rgb(223, 223, 223);;
-}
