+++ Dutify/src/components/cardsGrid/cards/addCard/addCard.jsx
@@ -1,7 +1,8 @@
 import React from "react";
 import { useReducer } from "react";
-import './addCardStyle.css';
 import { IoMdAddCircle } from "react-icons/io";
+import '../cardsStyle.css';
+import './addCardStyle.css';
 
 function PlaceHolder(state, action){
 
@@ -18,11 +19,11 @@ function AddCard({children}){
     const [state, dispatch] = useReducer(PlaceHolder , [])
 
     return (
-        <div className='card' id="main_container" 
+        <div className='card' id="card_container" 
             alt="Prueba">
-          <div className='d-flex flex-column align-items-center justify-content-center overlay fs-3'>
+          <div className='d-flex flex-column align-items-center justify-content-center cards-overlay fs-3'>
             <IoMdAddCircle className="addButton"/>
-            <p className="addCard-title">{children}</p>
+            <p className="addCard-title cards-title">{children}</p>
           </div>
         </div>
       );
