+++ Dutify/src/components/cardsGrid/cards/genreCard/genreStyle.css
