+++ Dutify/src/components/cardsGrid/cards/listCard/listCard.jsx
@@ -1,5 +1,6 @@
 import React, { useReducer} from "react";
 import "./listCardStyle.css";
+import "../cardsStyle.css";
 
 function PlaceHolder(state, action) {
   
@@ -17,13 +18,13 @@ function ListCard({ listName, background }) {
 
   return (
     <div
-      className="card"
-      id="listCard_container"
+      className="card "
+      id="card_container"
       alt={listName + "_Imagen"}
     >
-      <img src={background} className="card-img" />
-      <div className="listCard-overlay title-fade h-50 fs-2">
-        <p className="listCard-title">{listName}</p>
+      <img src={background} className="card-img"/>
+      <div className="cards-overlay title-fade">
+        <p className="cards-title">{listName}</p>
       </div>
     </div>
   );
